move websocket handling up to parent class
